[Refactor] Use assertLess for clearer assertion
In _assert_args_in_order, use self.assertLess to check if an argument is the last in a list.

Using assertLess is more semantically correct than assertNotEqual in this context. The original code checked if the index of an argument was not equal to the last possible index. However, the actual intent is to ensure that the index is less than the last index, which guarantees that there is at least one more element after it. Using assertLess makes this intent explicit and improves the readability of the test helper.

This change does not alter the existing logic because if an index is less than `len(arg_list) - 1`, it is also guaranteed to not be equal to `len(arg_list) - 1`. The behavior of the assertion remains the same.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I9120dcfd45e64aa083df5b7e3201e1aebef3501a