[Refactor] Use descriptive names for mock arguments
Rename unused mock arguments in  from , , and  to descriptive names prefixed with . This change improves code readability and maintainability by making it clear which mock object corresponds to which  decorator.

This refactoring does not alter the logic of the tests. The unused mock arguments are positional and their functionality is determined by the order of the decorators, not their names. By giving them descriptive names, we make the code easier for developers to understand without changing its behavior.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I17244808ed1dc0521711846cce392a17bfff878f