[Refactor] Improve robustness of a test helper
The  method in  could raise a  if the supplied  was not in , causing the test to error instead of fail.

This change adds a  check to ensure that  is in  before attempting to find its index. This makes the helper method more robust by providing a clearer failure message instead of an unhandled exception. This change does not alter any of the existing logic.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ic6f843348f399a97917121f671b811eec4b780de