[Refactor] Change HasValidTestMappingArgsTest base class
The HasValidTestMappingArgsTest class was inheriting from
AtestUnittestFixture, which sets up a pyfakefs fake filesystem.
However, the tests within this class do not perform any filesystem
operations, making the setup unnecessary overhead.

This change switches the base class to unittest.TestCase, which is
the standard base class for unittests and does not have the filesystem
setup overhead. This change is safe and does not alter the test logic
as the test does not depend on any functionality from AtestUnittestFixture.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ibca847db718d479ff0dbd231004c35f64d08d6b3