[Refactor] Move constant assignment out of loop in test
In , the variable  was being assigned the same constant value inside the test case loop.

This change moves the assignment outside the loop. This is a minor performance and readability improvement, as the constant is defined once.

This change does not alter the logic, as the value of  remains the same within the scope of the  call.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I90403133821b0669bc845cf66c95530f2ad50978