[Refactor] [minor bug fix] Avoid duplicate test discovery
Refactored the test discovery logic in `get_test_modules` to use a set to store the discovered test modules.

The original code used a list, which could lead to duplicate entries if both a `_unittest.py` and a `_unittest.pyc` file were present for the same module. This would cause the tests in that module to be run multiple times.

Using a set ensures that each test module is discovered and run only once, making the test discovery process more robust and efficient. This change does not alter the program's intended logic, but rather corrects a potential flaw in it.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I7a9cc0b2303adf222c5380bf0fe142a7ba03765d