[Refactor] Use f-string for string formatting in atest_utils_unittest.py
This change updates an old-style string formatting using the '%' operator to a modern f-string.
The f-string syntax is more readable and concise, especially for simple string interpolations.
This refactoring is purely stylistic and does not alter the existing logic or functionality of the test,
as the output string will remain identical.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ie08d7b60220afa5f10554b2c444926fdc345e20e