[Refactor] Fix indentation in atest_utils_unittest.py
This change improves the readability of the test_is_test_mapping_host_unit_test_only_specified function by correcting the indentation of the assertFalse arguments.
It does not alter the existing logic or behavior of the code in any way, as it's purely a formatting change.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I5888fa1a8db69d314cfde4bc3700174216ed1ec9