[Refactor] Refactor md5sum to read file in chunks
The `md5sum` function was reading the entire file content into memory. This can be problematic for large files, potentially leading to high memory usage. This change refactors the function to read the file in 4KB chunks, which is more memory-efficient. The logic of computing the MD5 hash remains the same.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ib37fec9ba250156eeb04ab461aacc2752b187069