[Refactor] Simplify dictionary update in get_android_junit_config_filters
Refactor `get_android_junit_config_filters` in `atest_utils.py` to use `setdefault` when updating the filter dictionary. This replaces the verbose pattern of `get`, `append`, and `update` with a more idiomatic and concise approach. This change does not alter the logic as `setdefault` handles both creating a new list if the key is missing and returning the existing list if present, which is equivalent to the previous implementation.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I94a111e2e3fbbeea1c35c43b1da8ee0f9b82c24e