[Refactor] Use pathlib for file handling in load_json_safely
Refactor `load_json_safely` to use `pathlib.Path` objects for file operations instead of `open()`.
This improves consistency with the rest of the file and leverages `pathlib` features.
It avoids instantiating `Path` object twice and uses `file_path.open()` which is cleaner.
This change does not alter the logic or functionality of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I6f9fd1c892363345320795a5ef774896ad75f892