[Refactor] Use f-string in save_md5
This change converts an old-style string formatting to an f-string in the save_md5 function. This improves readability and consistency with modern Python practices. The logic and behavior of the function remain unchanged as f-strings provide equivalent functionality to the .format() method but with a more concise syntax.

Test: atest atest_unittests

Flag: TEST_ONLY

OFF_PEAK_PRESUBMIT: true

Change-Id: I7d6183a182d8f99d1f4093d27bc0d1dfec3e067c