[Refactor] Simplify 'suites' assignment in _get_suite_to_modules
This change simplifies the assignment of the 'suites' variable in the
_get_suite_to_modules function. By using the 'or' operator, we eliminate
the redundant call to 'info.get("compatibility_suites")' and make the code
more concise and Pythonic. This modification preserves the existing logic
where 'suites' defaults to "['null-suite']" if 'compatibility_suites' is
missing or empty.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I8fb2069805c7401f8ef45c2577a126d04e88dcaf