[Refactor] Use f-string for 'No test found' message in CLITranslator
The change replaces a '.format()' call with an f-string for the 'No test found' log message in the '_find_test_infos' method of the CLITranslator class. This improves readability and aligns with modern Python style guidelines.

The original logic and functionality remain unchanged, as f-strings provide an equivalent way to format strings. Therefore, this change will not alter the existing behavior of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I31268a85088d18a9aa7f0f45410bed504cfa7679