[Refactor] Use set intersection for compatibility suites check
The original code used `any(x in ... for x in ...)` to check if any of the `TEST_MAPPING_SUITES` are present in the module's `compatibility_suites`. This refactoring uses `set(...).intersection(...)` which is a more standard and performant way to check for overlap between two collections in Python. It essentially performs the same logic: returning true if there is at least one common element.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ibaa252da84cef943f36093a93e5d43e7e45bbf95